+++ src/reports/Validation.ts
@@ -17,3 +17,19 @@ export type DependencyTypes = z.infer<typeof dependencyTypes>;
 export const TypeParameter = z.object({
     type: dependencyTypes
 });
+
+const httpString = z.custom<`http://${string}`>(value => {
+    if (typeof value === "string") return value.startsWith(`http://`);
+
+    return false;
+});
+
+const httpsString = z.custom<`https://${string}`>(value => {
+    if (typeof value === "string") return value.startsWith(`https://`);
+
+    return false;
+});
+
+export const urlType = z.union([httpString, httpsString]);
+
+export type Url = z.infer<typeof urlType>;
